requirements.txt
# Core dependencies
flask==2.3.3
werkzeug==2.3.7
PyPDF2
langchain==0.0.339
python-dotenv==1.0.0

# Additional dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
gunicorn==21.2.0
google-generativeai==0.3.1
langchain-community==0.0.10
pypdf==3.17.1
faiss-cpu==1.7.4
numpy==1.26.2
pydantic==2.5.2
langchain-google-genai==0.0.5
tqdm==4.66.1
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1